% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gallery.R
\name{gallery}
\alias{gallery}
\title{Overview of map projections}
\usage{
gallery(sample = 12, projections = NULL, ncol = 4)
}
\arguments{
\item{sample}{nb of maps to display (randomly)}

\item{projections}{vector of projection names}

\item{ncol}{ncol}
}
\value{
a base R plot.
}
\description{
Overview of map projections
}
\examples{
projs <- planisphere::registry(type = "core")
planisphere::gallery(projs)
}
