% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{planisphere}
\alias{planisphere-package}
\alias{planisphere}
\title{Package description}
\description{
A wrapper around D3.js for cartographic projections to design and visualize
world maps (planispheres). Projections are computed using D3.js spherical 
geometry rather than ellipsoidal geodesic models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://riatelab.github.io/planisphere/}
  \item Report bugs at \url{https://github.com/riatelab/planisphere/issues}
}

}
\author{
\strong{Maintainer}: Nicolas Lambert \email{nicolas.lambert@cnrs.fr} (\href{https://orcid.org/0000-0003-4976-6560}{ORCID})

Authors:
\itemize{
  \item Nicolas Lambert \email{nicolas.lambert@cnrs.fr} (\href{https://orcid.org/0000-0003-4976-6560}{ORCID})
}

}
