% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/registry.R
\name{registry}
\alias{registry}
\title{List available map projections}
\usage{
registry(type = NULL)
}
\arguments{
\item{type}{Character. Optional projection family. Must be one of:
"cylindrical", "conic", "azimuthal", "perspective",
"pseudocylindrical", "interrupted", "polyhedral",
"quincuncial", "regional", "pseudoazimuthal", "miscellaneous".}
}
\value{
A sorted unique character vector of projection names.
If `type` is provided, only projections belonging to the selected family
are returned.
}
\description{
Returns the list of available map projections registered in
\pkg{planisphere}. Projections are organized into thematic families
to facilitate exploration and selection.

By default, all projections are returned as a single sorted character vector.
Alternatively, a specific family of projections can be requested.
}
\examples{
# List all available projections

# List cylindrical projections
registry(type = "cylindrical")

# Explore regional projections
registry("regional")

# Visualize polyhedral projections
gallery(projections = registry("polyhedral"), verbose = FALSE)
}
